+++ v3/src/components/graph/plots/bar-chart/bar-chart.tsx
@@ -127,7 +127,7 @@ export const BarChart = observer(function BarChart({ abovePointsGroupRef, pixiPo
           })
         })
       })
-      renderBarCovers({ barCovers, barCoversRef, dataConfig, primaryAttrRole })
+      renderBarCovers({ barCovers, barCoversRef, graphModel, primaryAttrRole })
     }
 
     const anchor = circleAnchor