+++ v3/src/v2/codap-v2-data-set-types.ts
@@ -11,8 +11,10 @@ export interface ICodapV2Attribute {
   defaultMin?: number
   defaultMax?: number
   description?: string | null
+  // TODO_V2_IMPORT_EXPORT
   categoryMap?: any
   _categoryMap?: any
+  // TODO_V2_IMPORT_EXPORT
   colormap?: any
   _colormap?: any
   blockDisplayOfEmptyCategories?: boolean